Copy some (not all, yet) of the core files from Sage's libgap into gappy, unmodified.
The only difference is I've renamed util.pyx/pxd to core.pyx/pxd since
it has more than mere utilities, but is actually the core functionality
for initializing libgap.

This will not build or run yet at all, but the unmodified files are
added first so that their modifications for gappy can be easily tracked.