Minimal bits needed to get the existing Cython modules to actually build.
This includes a slightly non-trivial setup.py with build_ext class
capable of handling the Cython build, and handling the path to GAP_ROOT,
albeit also with the ability to compile against a "system-installed GAP"
on those distros where such a thing exists.

This still does not actually *work* as it depends on many runtime
dependencies from Sage.  The build-time cimports from Sage are
temporarily stubbed out with placeholders which will eventually be
removed.

In particular GapElement will need to be significantly reworked, and
possibly even renamed to avoid confusion with Sage's GapElement, as it
will not be taking advantage of Sage's Parent/Element framework and/or
coercion model (those bits will be added as wrappers in Sage's own
sage.libs.gap module, which this package does not aim to completely
replace).