Disable workspace handling and gaprc handling for now.
Although Sage has a notion of its own Sage-specific GAP workspace that
is auto saved/loaded we don't necessarily need this functionality for
gappy as workspace handling could be left as application-specific.

Some support for a "default" workspace much like Sage has could still
be incorporated, but as a later add-on.