Make the get_gap_memory_pool_size function work.  Adds psutil as a dependency.
Rather than having a separate _get_gap_memory_pool_size_MB as there was
in Sage, this get_gap_memory_pool_size always returns a value in some
units as parsed by the GAP CLI (with 'm' as the default).

The original implementation of _get_gap_memory_pool_size_MB always added
+1 to the value for reasons that aren't entirely clear.  Maybe it was
in order to round up?  But when the value is already an exact multiple
of megabytes there's no need for this.