Fix various direct usages of the global libgap instance in gappy.element.
It should only refer to the libgap instance actually passed in as
elements' parent.  For now there will still be only one libgap singleton
instance, since it's not possible (except possibly in the realm of
HPC-GAP) to have multiple GAP interpreters in a single process.  But we
should consider the possibility that that might be improved down the
line.

As noted, this change did probably break pickling of GapElements, which
will have to be revised later.

With this change it is now at least possible to *import* gappy.core
without any errors.  This does not mean that anything else is expected
to work yet, as there are still runtime dependencies on Sage and various
other broken pieces.