Add provisional support for running the doctests with pytest.
Requires pytest-cython, however the last release of pytest-cython is
very old and hasn't been updated for a while.  For the time being
a branch on my fork is needed to run this, but I am working on getting
upstream fixed:

https://github.com/lgpage/pytest-cython/pull/7

If I don't hear anything back from upstream I'll make a release under
a different name from my fork.

Note: With these changes it is now possible to find and run the
doctests, though not a single test actually passes yet, which is to be
expected.