Mass rename GapElement to GapObj.
This name was chosen so as to not confuse it with a subclass of Sage's
Element, though it does keep a _parent attribute and corresponding
parent() method like Sage, for now. It also emphasizes that it's a
wrapper around GAP Objs.

Sage's GapElement *might* be able to be implemented as a subclass of
GapObj and RingElement, though if that doesn't work (due to base class
incompatibility at the C level) GapElement can easily wrap GapObj.

The GapElement/GapObj subclasses are all also renamed with a more
PEP-8 compatibile name GapFoo instead of GapElement_Foo.

Remarkably, this enabled a small handful of (mostly trivial) tests to
pass.