Implement __add__ and other arithmetic special methods.
For now the Sage-specific _add_ and equivalents are kept, which may be
convenient later.  Sage's GapElement will want to use the _add_ versions
but override __add__ and the like with Sage's coercion model.

These just perform a very simple coercion to GapObj types where
possible.

Fixes a few tests, but some tests related to arithmetic are still broken
for some reason...