Partial attempt to fix handling of integers by restoring some code from the original GapElement_Integer.
Including note that large integers could be converted much more
efficiently; something to create an issue for later.

Note that large integers are still not fully working, because calling
gap(<bigint) is currently broken, as the original code did not
fully account for it when using Python ints (generally assuming a
bigint would be a Sage Integer).