New GapFunction.help() method which will also serve as the basis for providing docstrings for GapFunctions.
This works differently from and much better than the original approach
to GAP docstrings which was to call the GAP help facilities directly
using the `?<term>` syntax, which by default (but depending on user
settings!) dumps the results to the pager, and this is not so easy to
control without jumping through loops (which was actually my first
attempt, which involved registering a custom help viewer).

It turns out to be much easier to go directly through the
documented library interface to the GAP help system
(https://www.gap-system.org/Manuals/doc/ref/chap84.html).  There is
still a bit of unfortunate global state that needs to be managed to
get the desired output: This includes the GAPDocTextTheme, and the
current terminal screen width.

Additionally, the default behavior of the help system is to just open
a pager in the relevant chapter of the GAPDoc documentation for the
search term starting at the section which matches the search term, but
then the entire rest of the chapter is displayed starting from that
point.  While I can see how this is useful for interactive help, it's
not so useful if we want to generate a docstring for a function.  So
this uses the help API (at least in the case of GAPDoc; I'm not sure if
it works for other documentation formats) to find out what line of text
the next section starts on, and only extract up to that line, as the
doctest for GapFunction.help() demonstrates.