Add __doc__ as a property on GapFunctions which returns its help().
As noted, although the standard Python help() function does not display
instance docstrings, the IPython/Jupyter help with `<obj>?` *does*.

I think for the time being then, this is "good enough" for implementing
instancedocs on functions without any additional hacks.

Also caches the docstring once it's generated, since there is a
noticeable delay in generating it.