Replace most use of internal functions for strings with the public libgap API.  Refs #2.
The outlying exception currently is C_NEW_STRING which does not have an
exact analogue in the public libgap API.  The closest thing is
GAP_MakeString, but this assumes the string is NULL-terminated, meaning
it can't be used to convert arbitrary Python bytes objects containing
nulls.  In practice this won't often come up, but better safe than sorry
until a new API can be added.