Remove Gap.function_factory()
Between 0b02d9c0 and other changes the need for it seems to be pretty
well obviated.  If there's any use I can sort of think of, it's if
GAP global variables used internally by gappy are rebound by the user
it could break gappy.

But that was already the case if the variable is rebound before it was
possible to create it through function_factory.

The libgap interface in Sage could still provide this in its wrapper
but mark it deprecated.