Fix some compile-time warnings in GapList.__getitem__/__setitem__ about signed comparison.
At the same time add support for Python-style negative indexing.
Unfortunately this places a Python-specific limit on the allowed size
of GAP lists, but this limitation existed before this changes since
Cython was casting the index to a C long anyways.  This limitation
could be lifted in the remote chance it ever becomes a problem for
someone.