Prevent use of multiple Gap instances simultaneously.
In particular, if Gap() is used to create a new instance, it should not
be possible to keep using the default instance `gappy.gap`.

This is not my preferred approach; I would rather that the default
`gappy.gap` be a sort of singleton that always refers to the current
Gap instance.  This would be fine if I could make Gap a singleton class.
Maybe that's possible to do with a non-cdef wrapper.