Bug: Make sure GAP lists of characters are converted to strings.
GAP treats a list of characters equivalently to a string (it displays
it the same as a string, for example).  There was a workaround for this
in Sage which I removed because it didn't appear to have any impact
on the tests.

But in fact it did have a subtle impact that isn't obvious until
you display the type() of an element of such a list.  Before this
fix it was GapList, not GapString as expected.