Updates to handling conversion of built-in types.  In addition to Rational, adds support for numbers.Real and numbers.Integral.
Also makes it possible for converter functions (either `_gap_` or
converters registered with `register_converter` to return any of the
built-in types that gappy has built-in converters for.  E.g. to
convert, say, a custom "vector" type to GAP it's not necessary to write
the code to construct a GAP list.  Just returning the vector as a Python
tuple is sufficient.

Adds what I hope to be one of the few Sage-specific workarounds, for
how Sage misimplements the numbers.Rational ABC.