Simply the implementation of GapMethodProxy by making it actually wrap a GapFunction.
This works more the way Python methods work, in that it they also wrap
the function they are proxying.  In this way, GapFunction also works
more like Python functions work in that it implements __get__ to return
methods.

GapMethodProxy still subclasses GapObj, however, to make it more
immediately clear that it still wraps a GAP object.