Revert GapMethodProxy to being a subclass of GapFunction.
This is where we differ from Python, where the method type is *not*
a subclass of function.  But this is more consistent with the view
that GapMethodProxy still wraps a GAP function, and provides support
for GapFunction.__name__, GapFunction.help() etc. without having to
create additional wrappers.

It makes it seem almost redundant, then, to wrap the GapFunction, but
there's a reason for this: It makes it straightforward to work with
_GapLazyFunction as well, and its different __call__ implementation.