Cache lazy functions created with Gap.gap_function in an LRU cache.
Sage's Gap.function_factory also used an (unbounded) cache for its
created functions.

As a point of reference, there are less than 10 non-trivial uses of
Gap.function_factory (by which I mean, that define a new function
rather than just access an existing GAP built-in), so an LRU cache
of 128 should be more than enough for most usage in Sage.