[testing] Realized that this doctest was actually very broken, and it was only by some luck that it worked at all on Linux and Cygwin.
It did not work, however, on MacOS, and fixing it to work properly
actually demonstrates that it does not make a very good doctest (the
fact that it was a doctest at all is just because it was in Sage).

Moved it to a simpler regression test in tests/test_misc.py.