[bug] Fixes #12.
Kinda ugly workaround that unfortunately needs to reach for GAP
internals, though is effectively equivalent to the fix from
https://github.com/gap-system/gap/pull/4258 so when this fix is
available in new versions of GAP we can enable the workaround just
for older versions so long as they're supported.

Also probably need to switch to using sig_GAP_Enter() and
sig_GAP_Leave() in more places, GapFunction.__call__ is just the
most common.  But it should probably be used in almost all cases...?